GirafIntegrationTest/requirements.txt
requests>=2.23.0
pillow>=7.0.0